refactor: 컴포넌트 대문자로 수정 및 필요없는 파일 제거
