+++ src/utils/listing/toPostPropertyDetail.ts
